+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* [9297](https://github.com/osmosis-labs/osmosis/pull/9297) fix: unmatched exact amount out should result in error 
 
 ### State Breaking